Use BigDecimal for financial instrument prices

+++ src/main/java/com/stocktrading/dataloader1/remote/finnHub/FinancialInstrumentPriceDto.java
@@ -2,10 +2,11 @@ package com.stocktrading.dataloader1.remote.finnHub;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record FinancialInstrumentPriceDto(
-        @JsonProperty("p") Double price,
+        @JsonProperty("p") BigDecimal price,
         @JsonProperty("s") String symbol,
         @JsonProperty("t") Long timeStampUnixMili,
         @JsonProperty("v") Double volume,
